+++ webpack.config.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 
 const BUILD_DIR = path.resolve(__dirname, 'public');
@@ -19,6 +20,7 @@ module.exports = {
   },
 	plugins: [
 		HtmlWebpackPluginConfig,
+		new ExtractTextPlugin('styles.css'),
 		new CopyWebpackPlugin([
 			{ from: 'assets', to: 'assets' }
 		])
@@ -38,6 +40,10 @@ module.exports = {
         loader: 'babel-loader',
         exclude: /node_modules/
       },
+			{
+				test: /\.css$/,
+				loader: ExtractTextPlugin.extract('css-loader')
+			},
 			{
 				test: /\.(jpg|gif|png|svg|)$/,
 				loader: 'file-loader?name=[path][name].[ext]&context=./assets'